+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 ### Changelog
 
+**0.1.1**
+
+* Create automatic history app project (If you don't have)
+* Create multible History timeline add and remove feature is added
+* History item edit, change and remove feature is added
+* Selected timeline remove feature is added
+* Profile page has been added
+* Timeline items like-dislike feature is added
+
 **0.1.0**
 
 * Initial commit